Proxy object: whenever We call the object, object load from a another class is called proxy, default it is a lazy fetch, even though if the session closed all object details(values) get from a Proxy object
when we explicitly call the values of a collection it get from the
database not from the proxy object(default behaviour is Lazy loading)

Eager vs Lazy initialization: whenever we need to initialize a
collection inside of ohter object it does not pull the data of child
classes untill we need no need of this data is called lazy
initilization.

Eager Fetch: whenever we need child classes we can load those by
defining fetch=FetchType=EAGER

so that even though session closed all values(any collection) inside the
object will be available to proxy class.